agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bedrock-rag-query-service"
version = "0.1.0"
description = "Service for answering questions using AWS Bedrock RAG capabilities"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-aws>=0.2.22",
    "langchain-core>=0.3.56",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-mcp-tools",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "python-dotenv>=1.1.0",
    "pymongo>=4.6.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.284"
]

[tool.hatch.build.targets.wheel]
packages = ["."]
